Suppress all ZNC output in cron
ZNC produces an error message if one tries to relaunch it from cron. This is
unacceptable, so all of its output is suppressed for now. Later a proper test
should be added to launch ZNC only if it is running. The previous grep approach
failed because grep found the grep line every time, whether or not znc was
running. It also did not take into account the possibility of having several znc
instances running at the same time. These can distinguished from their "-d
<datadir>" option.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>